feat(82): dmn storage (#131)
Dmn engine now has persistence for DecisionDefinitions and has few queries to get them from db. ( tests included )
Sqlc overrides were removed for clarity. And for now types are cast from int64.
Fixed FindProcessDefinitionsById query since there were issues found with it during development.

* refactor: clean up bpmn persistance wrapper

* refactor: typo

* feat(82): dmn decisionDefinition storage

* feat(82): dmn decsionDefinition storage + refactor runtime pkg

refactor bpmnruntime, dmnruntime packages, remove int32 restrictions from DB

* feat(82): dmn decisionDefinition tests + refactor

* fix: sqlc generated code

* feat(82): remove dmn compression + refactor

* feat(82): final - with description

* fix: cleanup after mergin 82-dmn-storage with main

* refactor: dec cache configuration + file endings

* refactor: clean up bpmn persistance wrapper

---------

+++ internal/cluster/node.go
@@ -411,7 +411,7 @@ func (node *ZenNode) GetProcessDefinitions(ctx context.Context) ([]proto.Process
 	for _, def := range definitions {
 		resp = append(resp, proto.ProcessDefinition{
 			Key:        def.Key,
-			Version:    def.Version,
+			Version:    int32(def.Version),
 			ProcessId:  def.BpmnProcessID,
 			Definition: []byte(def.BpmnData),
 		})
@@ -431,7 +431,7 @@ func (node *ZenNode) GetLatestProcessDefinition(ctx context.Context, processId s
 	}
 	return proto.ProcessDefinition{
 		Key:        def.Key,
-		Version:    def.Version,
+		Version:    int32(def.Version),
 		ProcessId:  def.BpmnProcessID,
 		Definition: []byte(def.BpmnData),
 	}, nil
@@ -449,7 +449,7 @@ func (node *ZenNode) GetProcessDefinition(ctx context.Context, key int64) (proto
 	}
 	return proto.ProcessDefinition{
 		Key:        def.Key,
-		Version:    def.Version,
+		Version:    int32(def.Version),
 		ProcessId:  def.BpmnProcessID,
 		Definition: []byte(def.BpmnData),
 	}, nil
